requirements.txt
pywin32
DateTime
Kivy-Garden
MouseInfo
Pillow
PyAutoGUI
PyGetWindow
PyMsgBox
PyRect
PyScreeze
PyTweening
Pygments
SQLAlchemy
WMI
atomicwrites
attrs
certifi
chardet
colorama
comtypes
docutils
idna
importlib-metadata
keyboard
more-itertools
numpy
opencv-python
packaging
pandas
pip
pluggy
psutil
py
pymssql
pyodbc
pyparsing
pyperclip
pytesseract
pytest
python-dateutil
pytz
pywin32
pywinauto
requests
selenium
setuptools
six
urllib3
wcwidth
wheel
xlrd
zipp
zope.interface
openapi-codec
